requirements.txt
# Core Django and Web Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-allauth==0.57.0
django-grappelli==3.0.8

# Real-time Communication
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
mysqlclient==2.2.0
django-environ==0.11.2
dj-database-url==2.1.0

# Async Processing
celery==5.3.4
redis==5.0.1

# AI/ML Integration
torch==2.1.1
transformers==4.35.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1

# Geolocation and Maps
# django-geoposition==0.3.3  # Removed due to compatibility issues
geopy==2.4.0
folium==0.15.0

# File Handling and Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# Security and Authentication
PyJWT==2.8.0
cryptography==41.0.7

# Monitoring and Logging
sentry-sdk[django]==1.38.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3.post1